+++ app/assets/javascripts/Components/Result/image_viewer.jsx
@@ -7,19 +7,16 @@ export class ImageViewer extends React.Component {
     super();
   }
 
-  componentDidMount() {
-    if (this.props.url) {
-      this.ready_annotations();
-      this.props.annotations.forEach(this.display_annotation);
-    }
+  componentDidUpdate(prevProps, prevState) {
+    this.display_annotations();
   }
 
-  componentDidUpdate(prevProps, prevState) {
+  display_annotations = () => {
     if (this.props.url) {
       this.ready_annotations();
       this.props.annotations.forEach(this.display_annotation);
     }
-  }
+  };
 
   ready_annotations = () => {
     window.annotation_type = ANNOTATION_TYPES.IMAGE;
@@ -32,15 +29,6 @@ export class ImageViewer extends React.Component {
       !this.props.released_to_students
     );
   };
-
-  componentWillUnmount() {
-    let box = document.getElementById('sel_box');
-    if (box) {
-      box.style.display = 'none';
-      box.style.width   = '0';
-      box.style.height  = '0';
-    }
-  }
 
   display_annotation = (annotation) => {
     add_annotation_text(annotation.annotation_text_id, annotation.content);
@@ -56,8 +44,10 @@ export class ImageViewer extends React.Component {
   render() {
     return (
       <div id='image_container' className='image_container'>
+        <div key='sel_box' id='sel_box' className='annotation-holder-active' style={{display: 'none'}}/>
         <img id='image_preview'
           src={this.props.url}
+          onLoad={this.display_annotations}
           alt={I18n.t('results.cant_display_image')} />
       </div>
     );