+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## [v1.9.4]
 - Fixed bug where bonuses and deductions were not displayed properly (#4699)
+- Fixed bug where image annotations did not stay fixed relative to the image (#4706)
+- Fixed bug where image annotations did not load properly (#4706)
 
 ## [v1.9.3]
 - Fixed inverse association bug with assignments (#4551)